+++ packages/access-policy/package.json
@@ -29,6 +29,7 @@
   },
   "dependencies": {
     "@tupaia/constants": "workspace:*",
+    "@tupaia/tsutils": "workspace:*",
     "@tupaia/utils": "workspace:*"
   }
 }
