+++ packages/central-server/src/permissions/constants.js
@@ -1,3 +1,6 @@
-export const BES_ADMIN_PERMISSION_GROUP = 'BES Admin';
-export const TUPAIA_ADMIN_PANEL_PERMISSION_GROUP = 'Tupaia Admin Panel';
-export const VIZ_BUILDER_PERMISSION_GROUP = 'Viz Builder User';
+/* Re-export for backward compatibility. Prefer importing directly from @tupaia/constants. */
+export {
+  BES_ADMIN_PERMISSION_GROUP,
+  TUPAIA_ADMIN_PANEL_PERMISSION_GROUP,
+  VIZ_BUILDER_PERMISSION_GROUP,
+} from '@tupaia/constants';
