+++ packages/constants/src/permission.ts
@@ -18,3 +18,5 @@ export const API_CLIENT_PERMISSIONS = [
 ] as const;
 
 export const BES_ADMIN_PERMISSION_GROUP = 'BES Admin';
+export const TUPAIA_ADMIN_PANEL_PERMISSION_GROUP = 'Tupaia Admin Panel';
+export const VIZ_BUILDER_PERMISSION_GROUP = 'Viz Builder User';
